requirements.txt
aioreloader==0.4.0
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
asyncpg==0.25.0
click==8.0.4
databases==0.5.5
dill==0.3.4
environs==9.5.0
fastapi==0.75.0
funcy==1.17
greenlet==1.1.2
gunicorn==20.1.0
h11==0.11.0
idna==3.3
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
psycopg2-binary==2.9.3
pydantic==1.9.0
pyfunctional==1.4.3
pyparsing==3.0.7
python-dotenv==0.19.2
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
structlog==21.5.0
tabulate==0.8.9
typing_extensions==4.1.1
uvicorn==0.17.6
uvloop==0.16.0
httpx~=0.22.0